+++ include/timer.h
@@ -1,11 +1,13 @@
 #ifndef TIMER_H
 #define TIMER_H
 
+#include <types.h>
+
 /* Timers */
 // Overo uses 26MHz system clock
 #define TICKS_PER_MSEC 26000ULL
 #define TICKS_PER_USEC (TICKS_PER_MSEC/1000)
-unsigned long long read_timer(void);
+uint64_t read_timer(void);
 void udelay(int usec);
 
 #endif /* TIMER_H */
